+++ R/describe.R
@@ -72,11 +72,13 @@ NULL
 ##'
 ##' @param object Either an instance of class `Matrix`, [PSM()] or
 ##'     [ConnectedComponents()].
-describeProteins <- function(object) {
+##'     
+##' @param ... Additional arguments passed to [makeAdjacencyMatrix()].
+describeProteins <- function(object, ...) {
     if (is(object, "PSM")) {
-        adj <- makeAdjacencyMatrix(object)
+        adj <- makeAdjacencyMatrix(object, ...)
     } else if (is(object, "ConnectedComponents")) {
-        adj <- adjacencyMatrix(object)
+        adj <- adjacencyMatrix(object, ...)
     } else if (is(object, "Matrix")) {
         adj <- object
     } else stop("Object must be of class 'Matrix', 'PSM' or 'ConnectedComponents'")
@@ -103,11 +105,11 @@ describeProteins <- function(object) {
 ##' @export
 ##'
 ##' @rdname describeProteins
-describePeptides <- function(object) {
+describePeptides <- function(object, ...) {
     if (is(object, "PSM")) {
-        adj <- makeAdjacencyMatrix(object)
+        adj <- makeAdjacencyMatrix(object, ...)
     } else if (is(object, "ConnectedComponents")) {
-        adj <- adjacencyMatrix(object)
+        adj <- adjacencyMatrix(object, ...)
     } else if (is(object, "Matrix")) {
         adj <- object
     } else stop("Object must be of class 'Matrix', 'PSM' or 'ConnectedComponents'")