+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## PSMatch 1.13.2
 
+- Add possibility to split protein groups in PSM data with `makeAdjacencyMatrix`
+(see [issue #30](https://github.com/rformassspectrometry/PSMatch/issues/30))
 - Update Fragments vignette. 
 - Improve `labelFragments()`runtime 
 (see [issue #25](https://github.com/rformassspectrometry/PSMatch/issues/25)).